Move delete task to revisionManager
