+++ lib/registries/addon/edit-revision/template.hbs
@@ -85,7 +85,6 @@
                         <div local-class='RightWrapper'>
                             <div local-class='RightNavWrapper'>
                                 <EditRevision::-Components::RightNav
-                                    @onDelete={{this.deleteRevision}}
                                     @revisionManager={{revisionManager}}
                                     @navManager={{navManager}}
                                 />
